#1687 rename ereignismeldung to ergebnismeldung
