+++ wls-gui-wahllokalsystem/src/types/ergebnismeldung/BezirkUndWahlID.ts
